refactor(runbotics-desktop): :zap: Refactor after CR
